:hammer: [Fix] 유저 밴 기능 1차 수정 #95
